backend/requirements.txt
web3
Flask
pytest
py-solc-x